eckey: move eckey functions to their own files; add docs
